Fix CoreRT build breaks
