+++ src/System.Private.CoreLib/src/System.Private.CoreLib.csproj
@@ -194,9 +194,8 @@
     <Compile Include="System\AppContextConfigHelper.cs" />
     <Compile Include="System\Buffer.cs" />
     <Compile Include="System\Collections\Generic\ArraySortHelper.CoreRT.cs" />
-    <Compile Include="System\Collections\Generic\Comparer.cs" />
-    <Compile Include="System\Collections\Generic\CompatibilityEqualityComparers.cs" />
-    <Compile Include="System\Collections\Generic\EqualityComparer.cs" />
+    <Compile Include="System\Collections\Generic\Comparer.CoreRT.cs" />
+    <Compile Include="System\Collections\Generic\EqualityComparer.CoreRT.cs" />
     <Compile Include="System\Collections\Generic\EqualOnlyComparer.cs" />
     <Compile Include="System\DefaultBinder.CanConvert.cs" />
     <Compile Include="System\Diagnostics\Contracts\ContractsBCL.cs" />
